webstorage/iframe/local_security_iframe.html

<!DOCTYPE HTML>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(function() {
  var ls = parent.localStorage;
  document.domain = document.domain;
  assert_throws("SecurityError", function() {
    ls.clear();
  });
});
</script>
